MDL-63120 - Example of one way to solve
There is a bug in the Moodle code that involves MariaDB/MySQL when
trying to join > 60 tables.

If the number of activities needed to earn a badge is greater than
60 (one other join is included in the SQL already), then awarding
the badge fails.

This solution works, but may be considered "clunky". I'm not a PHP
developer, so feel free to give feedback.

In badges/criteria/award_criteria.php, the abstract function
get_completed_criteria_sql() is defined, and then implemented by
  *award_criteria.php
  *award_critieria_activty.php
  *award_critieria_badge.php
  *award_critieria_cohort.php
  *award_criteria_competency.php
  *award_criteria_course.php
  *award_criteria_courseset.php
  *award_criteria_manual.php
  *award_criteria_overall.php
  *award_criteria_profile.php

This function returns an array list with the 'join', the 'where',
and the 'params' for the SQL. This solution returns a multi-demensional
array, splitting up the join/where/params into groups of 58 table joins.

The code in badge.php then iterates over the arrays, ensuring that we do
not hit the 61 table limit.